feat: rules-based automatic parameterisation from pathnames
There are multiple ways to automatically parameterise pathnames. This is one approach, and generally the fastest, a set of functions each accept the pathname of an incoming request and determine which parts are parameters. At present, only UUIDs are identified. Other approaches might include custom regex strings, or custom function logic.

- Ability to automatically identify path parameters from a set of rules
- Designed to be used in conjunction with more computationally expensive methods
- Rules are functions. New functions can be added to identify path parameters from string patterns
- The only current rule is that UUIDs in pathnames will be parameterised